+++ src/main/java/algorithms/project/benchmark/HCE.java
@@ -6,8 +6,8 @@ package algorithms.project.benchmark;
 public class HCE implements Benchmark {
 
     @Override
-    public float benchmark(Float... params) {
-        float fitness = 0;
+    public Double benchmark(Double... params) {
+        double fitness = 0;
 
         for (int i = 0; i < params.length; i++) {
             fitness += Math.pow(1000000, (i - 1) / (params.length - 1)) * params[i] * params[i];
